move profunctor and strong to arrow
