+++ core/src/main/scala/cats/data/IndexedReaderWriterStateT.scala
@@ -1,7 +1,8 @@
 package cats
 package data
 
-import cats.functor.{ Bifunctor, Contravariant, Profunctor, Strong }
+import cats.arrow.{Profunctor, Strong}
+import cats.functor.{Bifunctor, Contravariant}
 import cats.syntax.either._
 
 /**
