+++ core/src/main/scala/cats/syntax/strong.scala
@@ -1,6 +1,6 @@
 package cats
 package syntax
 
-import cats.functor.Strong
+import cats.arrow.Strong
 
 trait StrongSyntax extends Strong.ToStrongOps
